Optimization of the OpenGL test interface

+++ xflr5/xflr5v6/xfl3d/resources/shaders/fractal/fractal_FS.glsl
@@ -1,4 +1,4 @@
-#version 410 core
+#version 330
 
 uniform int maxiters;
 uniform float maxlength;
@@ -13,16 +13,16 @@ void main(void)
 {
 //    The Mandelbrot set is the set of complex numbers c for which the function f_c(z) = z²+c
 //    does not diverge when iterated from z=0
-    double ptx, pty, ptx2, pty2;
-    double tmpx;
+    float ptx, pty, ptx2, pty2;
+    float tmpx;
     int iter=0;
     do
     {
         ptx2 = ptx*ptx;
         pty2 = pty*pty;
         tmpx = ptx;
-        ptx = ptx2 - pty2  + double(posx);
-        pty = 2.0*tmpx*pty + double(posy);
+        ptx = ptx2 - pty2  + float(posx);
+        pty = 2.0*tmpx*pty + float(posy);
         iter++;
     }
     while(ptx2 + pty2<maxlength*maxlength && iter<maxiters);
